refactor(lwm): extract Large Mover ledger id parser
