+++ apps/ledger-live-mobile/src/mvvm/features/LandingPages/screens/LargeMoverLandingPage/utils/parseLargeMoverLedgerIds.ts
@@ -1,0 +1,7 @@
+export function parseLargeMoverLedgerIds(ledgerIds: string): string[] {
+  const ids = ledgerIds
+    .split(",")
+    .map(id => id.trim().toLowerCase())
+    .filter(Boolean);
+  return [...new Set(ids)];
+}
